+++ app/main/preload.js
@@ -5,9 +5,9 @@ import _http from 'http';
 import { ipcRenderer as _ipcRenderer, remote as _remote } from 'electron';
 import _electronLog from 'electron-log';
 import ElectronStore from 'electron-store';
+import config from 'app/config';
 import { environment } from './environment';
 
-const { isDev } = environment;
 const _process = process;
 const _electronStore = new ElectronStore();
 
@@ -51,7 +51,7 @@ process.once('loaded', () => {
     // $FlowFixMe
     global.spectronRequire = __non_webpack_require__; // eslint-disable-line
   }
-  if (!isDev) {
+  if (!config.isDev) {
     // ESLint will warn about any use of eval(), even this one
     // eslint-disable-next-line
     global.eval = () => {