+++ app/renderer/api/api.js
@@ -12,6 +12,7 @@ import { Keyring } from '@polkadot/keyring';
 
 import { generateMnemonic } from 'renderer/utils/crypto';
 import { stringifyData, stringifyError } from 'common/utils/logging';
+import MNEMONIC_RULE from 'renderer/constants/mnemonic';
 import { getSystemHealth } from './nodes/requests/getSystemHealth';
 import { Logger } from '../utils/logging';
 
@@ -75,7 +76,7 @@ const toyKeyring = toyKeyringFromNames([
   'Courtney',
   'Drew',
   'Emily',
-  'Frank'
+  'Frank',
 ]);
 
 export default class CennzApi {
@@ -113,7 +114,7 @@ export default class CennzApi {
     await generatePaperWalletChannel.send({
       address: request.address,
       filePath,
-      mnemonics: request.mnemonic.split(', '),
+      mnemonics: request.mnemonic.split(MNEMONIC_RULE),
       isMainnet: true,
       networkName: request.networkName,
       buildLabel,
@@ -146,15 +147,25 @@ export default class CennzApi {
       console.log(`walletAddresses: ${walletAddresses}`);
       const defaultAccountPublicAddress = walletAddresses[0];
       resultWallet.defaultAccountPublicAddress = defaultAccountPublicAddress;
-      console.log(`resultWallet.defaultAccountPublicAddress: ${resultWallet.defaultAccountPublicAddress}`);
+      console.log(
+        `resultWallet.defaultAccountPublicAddress: ${resultWallet.defaultAccountPublicAddress}`
+      );
 
-      const stakingTokenFreeBalance = await window.odin.api.cennz.getGenericAssetFreeBalance(PreDefinedAssetIdObj.STAKING_TOKEN.BN, resultWallet.defaultAccountPublicAddress);
+      const stakingTokenFreeBalance = await window.odin.api.cennz.getGenericAssetFreeBalance(
+        PreDefinedAssetIdObj.STAKING_TOKEN.BN,
+        resultWallet.defaultAccountPublicAddress
+      );
       resultWallet.stakingTokenFreeBalance = stakingTokenFreeBalance.toString(10);
       console.log(`resultWallet.stakingTokenFreeBalance: ${resultWallet.stakingTokenFreeBalance}`);
 
-      const spendingTokenFreeBalance = await window.odin.api.cennz.getGenericAssetFreeBalance(PreDefinedAssetIdObj.SPENDING_TOKEN.BN, resultWallet.defaultAccountPublicAddress);
+      const spendingTokenFreeBalance = await window.odin.api.cennz.getGenericAssetFreeBalance(
+        PreDefinedAssetIdObj.SPENDING_TOKEN.BN,
+        resultWallet.defaultAccountPublicAddress
+      );
       resultWallet.spendingTokenFreeBalance = spendingTokenFreeBalance.toString(10);
-      console.log(`resultWallet.spendingTokenFreeBalance: ${resultWallet.spendingTokenFreeBalance}`);
+      console.log(
+        `resultWallet.spendingTokenFreeBalance: ${resultWallet.spendingTokenFreeBalance}`
+      );
 
       const genericAssetNextAssetId = await window.odin.api.cennz.getGenericAssetNextAssetId();
       resultWallet.genericAssetNextAssetId = genericAssetNextAssetId.toString(10);
@@ -162,23 +173,32 @@ export default class CennzApi {
 
       // extract data from wallet to accounts
       const accounts = new Map();
-      for(const walletAddress of walletAddresses) {
+      for (const walletAddress of walletAddresses) {
         console.log(`walletAddress: ${walletAddress}`);
 
         const assets = new Map();
         // eslint-disable-next-line no-await-in-loop
-        const stakingTokenAsset = await this.getCennznetWalletAsset(PreDefinedAssetIdObj.STAKING_TOKEN.BN, walletAddress);
+        const stakingTokenAsset = await this.getCennznetWalletAsset(
+          PreDefinedAssetIdObj.STAKING_TOKEN.BN,
+          walletAddress
+        );
         assets[PreDefinedAssetIdObj.STAKING_TOKEN.BN] = stakingTokenAsset;
 
         // eslint-disable-next-line no-await-in-loop
-        const spendingTokenAsset = await this.getCennznetWalletAsset(PreDefinedAssetIdObj.SPENDING_TOKEN.BN, walletAddress);
+        const spendingTokenAsset = await this.getCennznetWalletAsset(
+          PreDefinedAssetIdObj.SPENDING_TOKEN.BN,
+          walletAddress
+        );
         assets[PreDefinedAssetIdObj.SPENDING_TOKEN.BN] = spendingTokenAsset;
 
         // TODO base on nextAssetId fetch all custom token ?
-        for(const customToken of CustomTokenAssetId) {
+        for (const customToken of CustomTokenAssetId) {
           const customTokenAssetIdAsBN = new BN(customToken, 10);
           // eslint-disable-next-line no-await-in-loop
-          const customTokenAsset = await this.getCennznetWalletAsset(customTokenAssetIdAsBN, walletAddress);
+          const customTokenAsset = await this.getCennznetWalletAsset(
+            customTokenAssetIdAsBN,
+            walletAddress
+          );
           assets[customTokenAssetIdAsBN] = customTokenAsset;
         }
 
@@ -202,15 +222,24 @@ export default class CennzApi {
     }
   };
 
-  getCennznetWalletAsset = async (assetId: BN, walletAddress: String): Promise<CennznetWalletAsset> => {
+  getCennznetWalletAsset = async (
+    assetId: BN,
+    walletAddress: String
+  ): Promise<CennznetWalletAsset> => {
     Logger.debug('CennznetApi::getCennznetWalletAsset called');
     try {
-      const freeBalanceBN = await window.odin.api.cennz.getGenericAssetFreeBalance(assetId, walletAddress);
-      const freeBalance = {toBN: freeBalanceBN, toString: freeBalanceBN.toString(10)};
-      const reservedBalanceBN = await window.odin.api.cennz.getGenericAssetReservedBalance(assetId, walletAddress);
-      const reservedBalance = {toBN: reservedBalanceBN, toString: reservedBalanceBN.toString(10)};
+      const freeBalanceBN = await window.odin.api.cennz.getGenericAssetFreeBalance(
+        assetId,
+        walletAddress
+      );
+      const freeBalance = { toBN: freeBalanceBN, toString: freeBalanceBN.toString(10) };
+      const reservedBalanceBN = await window.odin.api.cennz.getGenericAssetReservedBalance(
+        assetId,
+        walletAddress
+      );
+      const reservedBalance = { toBN: reservedBalanceBN, toString: reservedBalanceBN.toString(10) };
       const totalBalanceBN = freeBalanceBN.add(reservedBalanceBN);
-      const totalBalance = {toBN: totalBalanceBN, toString: totalBalanceBN.toString(10)};
+      const totalBalance = { toBN: totalBalanceBN, toString: totalBalanceBN.toString(10) };
       const asset = new CennznetWalletAsset({
         assetId,
         address: walletAddress,
@@ -256,12 +285,14 @@ export default class CennzApi {
       const wallet = new Wallet();
       await wallet.createNewVault(request.passphrase);
       await wallet.addAccount();
-      const backup = await wallet.export(request.passphrase);
+      const exportedWallet = await wallet.export(request.passphrase);
+      const mnemonic = exportedWallet && exportedWallet[0].data.mnemonic;
 
       const cennznetWallet = new CennznetWallet({
         id: uuid(),
         name: request.name,
         hasPassword: request.passphrase !== null,
+        mnemonic,
         wallet,
       });
       Logger.debug('CennznetApi::createWalletWithHDKeyRing success');
@@ -363,8 +394,9 @@ export default class CennzApi {
     Logger.debug('CennznetApi::getGenericAssetFreeBalance called');
     try {
       const freeBalance = await this.ga.getFreeBalance(assetId, walletAddress);
-      console.log(freeBalance);
-      Logger.debug(`CennznetApi::getGenericAssetFreeBalance freeBalance: ${freeBalance}, ${typeof freeBalance}`);
+      Logger.debug(
+        `CennznetApi::getGenericAssetFreeBalance freeBalance: ${freeBalance}, ${typeof freeBalance}`
+      );
       return new BN(freeBalance.toString(10), 10);
     } catch (error) {
       Logger.error('CennznetApi::getGenericAssetFreeBalance error: ' + stringifyError(error));
@@ -407,36 +439,45 @@ export default class CennzApi {
     }
   };
 
-
   /**
    * @param assetId
    * @param fromWalletAddress
    * @param toWalletAddress
    * @param amount
    * @param wallet
-   * @returns {Promise<Object>}
+   * @returns {Promise<Hash>}
    */
-  doGenericAssetTransfer = async (assetId: BN,  fromWalletAddress: string, toWalletAddress: string, amount: BN, wallet): Promise<Object> => {
+  doGenericAssetTransfer = async (assetId: BN,  fromWalletAddress: string, toWalletAddress: string, amount: BN, wallet: CennznetWallet): Promise<Hash> => {
     Logger.debug('CennznetApi::doGenericAssetTransfer called');
+    Logger.debug(`assetId: ${assetId.toString(10)}, amount: ${amount.toString(10)}, fromWalletAddress: ${fromWalletAddress}, toWalletAddress: ${toWalletAddress}`);
     try {
-      await this.api.setSigner(wallet);
-      const tx = await this.ga.transfer(assetId, fromWalletAddress, amount).send({from: toWalletAddress});
-      // const tx = await this.api.tx.genericAsset.transfer(assetId, fromWalletAddress, amount).send({from: toWalletAddress});
-      // this.api.setSigner(null);
-      // const tx = await new Promise(resolve => {
-      //   let submittableExtrinsic = this.api.tx.genericAsset.transfer(assetId, toWalletAddress, amount);
-      //   debugger;
-      //   submittableExtrinsic.signAndSend(toyKeyring.bob, res => {
-      //     if (res.status.type === 'Finalised') {
-      //       console.log('tx finalized');
-      //       resolve();
-      //     }
-      //   });
+
+      // reload wallet object
+      const originalWallet = new Wallet({
+        vault: wallet.wallet.vault,
+        keyringTypes: [HDKeyring, SimpleKeyring], // add keyringTypes: [HDKeyring, SimpleKeyring] if SimpleKeyring is used
+      });
+      await originalWallet.unlock(''); // TODO switch to pin code
+      Logger.debug('unlock');
+
+      this.api.setSigner(originalWallet);
+      Logger.debug('setSigner');
+
+      // const tx = await this.ga.transfer(assetId, toWalletAddress, amount).send({from: fromWalletAddress}, async (status) => {
+      //   console.log('transfer status');
+      //   console.log(status);
+      //   if (status.type === 'Finalised' && status.events !== undefined) {
+      //     console.log('success');
+      //     const balance: BN = await this.ga.getFreeBalance(assetId, toWalletAddress);
+      //     console.log('balance');
+      //     console.log(balance);
+      //   }s
       // });
-      Logger.debug(`CennznetApi::genericAssetTransfer tx: ${tx}`);
-      return tx;
+      const txHash = await this.ga.transfer(assetId, toWalletAddress, amount).send({from: fromWalletAddress});
+      Logger.debug(`CennznetApi::doGenericAssetTransfer txHash ${txHash}`);
+      return txHash;
     } catch (error) {
-      Logger.error('CennznetApi::genericAssetTransfer error: ' + stringifyError(error));
+      Logger.error('CennznetApi::doGenericAssetTransfer error: ' + stringifyError(error));
       throw new GenericApiError();
     }
   };