+++ app/renderer/components/layout/TopBar/index.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 import SVGInline from 'react-svg-inline';
+
+import { NetworkNameMapping } from 'common/types/cennznet-node.types';
 import { colors } from 'renderer/theme';
 import logoImg from 'renderer/assets/img/centrality-logo.svg';
 import { NETWORK_OPTIONS, getNetworkOptionPair } from 'renderer/pages/chooseNetworkPage';
@@ -8,6 +10,9 @@ import { Select } from 'components';
 import SwitchNetworkWarningModal from './TopBarWarningModal';
 import UploadGenesisFileModal from './UploadGenesisModal';
 import withContainer from './TopBarContainer';
+import { environment } from '../../../../main/environment';
+
+const  { isDevOrDebugProd } = environment;
 
 const Wrapper = styled.div`
   display: flex;
@@ -36,6 +41,37 @@ const NetworkSectionWrapper = styled.div`
   min-width: 12rem;
 `;
 
+const InfoWrapper = styled.div`
+  display: flex;
+`;
+
+const InfoValue = styled.div`
+  height: 16px;	
+  color: ${colors.N0};
+  line-height: 21px;	
+  text-align: right;
+`;
+
+const InfoDesc = styled.div`
+  height: 16px;	
+  color: ${colors.textMuted};
+  line-height: 21px;	
+  text-align: right;
+`;
+
+const DevInfo = ({isSynced, syncPercentage, blockNum, blockSpeed }) => {
+  if(!isDevOrDebugProd) {
+    return <div />;
+  }
+  return (
+    <div>
+      <div>{isSynced ? '100% ' : syncPercentage + ' '} synced</div>
+      <div>(block height: {blockNum})</div>
+      <div>(block speed: {blockSpeed})</div>
+    </div>
+  );
+};
+
 const TopBar = ({
   nodeSystem,
   remoteStream,
@@ -58,8 +94,7 @@ const TopBar = ({
   const { blockNum: remoteBlockNum, bps: remoteBps } = syncRemoteStream;
   const { blockNum: localBlockNum, bps: localBps } = syncStream;
   const blockNum = `#${localBlockNum} / #${remoteBlockNum}`;
-  const blockSpeed =
-    localBps && remoteBps ? `${localBps.toFixed(2)}bps / ${remoteBps.toFixed(2)}bps` : '0 bps';
+  const blockSpeed = `${localBps?localBps.toFixed(2):0}bps / ${remoteBps?remoteBps.toFixed(2):0}bps`;
 
   const percentage = remoteBlockNum > 0 ? ((localBlockNum / remoteBlockNum) * 100).toFixed(2) : 0;
   const syncPercentage = `${percentage}%`;
@@ -71,10 +106,11 @@ const TopBar = ({
         <NetworkSectionContainer>
           <NetworkSectionWrapper>
             <Select
+              backgroundColor={colors.V800}
               fontWeight="600"
               fontSize="16px"
               borderColor="transparent"
-              value={getNetworkOptionPair(networkName, 'label')}
+              value={getNetworkOptionPair(NetworkNameMapping[networkName], 'value')}
               onChange={selected => {
                 setIsOpenNetworkWarningModal(true);
                 setSelectedNetwork(selected);
@@ -83,10 +119,17 @@ const TopBar = ({
             />
           </NetworkSectionWrapper>
         </NetworkSectionContainer>
-        <div>
-          <div>{isSynced ? '100% ' : syncPercentage + ' '} synced</div>
-          <div>(block speed: {blockSpeed})</div>
-        </div>
+        <DevInfo {...{isSynced, syncPercentage, blockNum, blockSpeed}} />
+        <InfoWrapper>
+          <div style={{marginRight: '2rem'}}>
+            <InfoValue>{localBlockNum || 0}/{remoteBlockNum || 0}({syncPercentage})</InfoValue>
+            <InfoDesc>block height</InfoDesc>
+          </div>
+          <div>
+            <InfoValue>{localBps || 0}</InfoValue>
+            <InfoDesc>bps</InfoDesc>
+          </div>
+        </InfoWrapper>
       </HeaderWrapper>
       <SwitchNetworkWarningModal
         setIsOpenNetworkWarningModal={setIsOpenNetworkWarningModal}