+++ app/renderer/pages/chooseNetworkPage/index.js
@@ -7,6 +7,9 @@ import { Button, FileUploader, Select, PageHeading } from 'components';
 import { Logger } from 'renderer/utils/logging';
 import { NetworkNameOptions } from 'common/types/cennznet-node.types';
 import withContainer from './container';
+import { environment } from '../../../main/environment';
+
+const { isDevOrDebugProd } = environment;
 
 const ChooseNetworkWrapper = styled.div`
   width: 60%;
@@ -46,13 +49,17 @@ const ButtonWrapper = styled.div`
 export const NETWORK_OPTIONS = [
   // { label: 'CENNZnet DEV(OLD)', value: NetworkNameOptions.CENNZNET_DEV },
   // { label: 'CENNZnet UAT(OLD)', value: NetworkNameOptions.CENNZNET_UAT },
+  { label: 'CENNZnet RIMU(UAT)', value: NetworkNameOptions.CENNZNET_RIMU }, // TODO should we add *-latest runtime option?
   { label: 'CENNZnet KAURI(DEV)', value: NetworkNameOptions.CENNZNET_KAURI },
-  { label: 'CENNZnet RIMU(UAT)', value: NetworkNameOptions.CENNZNET_RIMU },// TODO should we add *-latest runtime option?
-  { label: 'Local test net', value: NetworkNameOptions.LOCAL_TESTNET }, // TODO hide for non-DEV mode
 ];
 
-export const getNetworkOptionPair = (value, param = 'value') =>
-  NETWORK_OPTIONS.find(option => option[param] === value);
+if (isDevOrDebugProd) {
+  NETWORK_OPTIONS.push({ label: 'Local test net', value: NetworkNameOptions.LOCAL_TESTNET });
+}
+
+export const getNetworkOptionPair = (value, param = 'value') => {
+  return NETWORK_OPTIONS.find(option => option[param] === value);
+};
 
 const ChooseNetWork = ({
   selectedNetwork,
@@ -79,6 +86,7 @@ const ChooseNetWork = ({
             <div>Choose network</div>
             <NetworkOptionWrapper>
               <Select
+                backgroundColor={colors.V800}
                 value={selectedNetwork}
                 onChange={selected => {
                   Logger.info('selected value', selected);