+++ app/renderer/pages/wallet/TextField.js
@@ -6,41 +6,53 @@ import { colors } from 'renderer/theme';
 
 const Label = styled.label`
   margin: 0.5rem 0.5rem 0 0;
-  color: ${colors.N200};
+  color: ${colors.N0};
+  font-size: 16px;	
+  font-weight: 600;	
+  line-height: 22px;
 `;
 
 const Wrapper = styled.div`
-  display: flex;
-  align-items: center;
   margin-bottom: 0.5rem;
-
-  input {
-    width: 10rem;
-  }
 `;
 
-const WorldField = ({
+const ErrorMessage = styled.div`
+  margin-top: 0.5rem;
+  height: 1.2rem;
+  color: ${colors.danger};
+`;
+
+const ErrorMessagePleaceHolder = styled.div`
+  margin-top: 0.5rem;
+  height: 1.2rem;
+`;
+
+const TextField = ({
   field, // Formik prop - { name, value, onChange, onBlur }
   form, // Formik prop - { touched, errors }, also values, setXXXX, handleXXXX, dirty, isValid, status, etc.
   labelText,
   placeholder,
   noTickShow,
+  width,
+  minWidth,
+  readOnly = false,
 }) => {
   const { name, value, onChange, onBlur } = field;
   const { touched, errors } = form;
   const fieldError = R.path(R.split('.')(name))(errors);
   const fieldTouched = R.path(R.split('.')(name))(touched);
-  const recomposedValue = value && value.toLowerCase().trim();
   return (
-    <Wrapper>
+    <Wrapper style={{minWidth, width}}>
       {labelText && <Label htmlFor={name}>{labelText}</Label>}
       <Input
-        value={recomposedValue || ''}
+        value={value || ''}
         {...{ placeholder, name, onChange, onBlur }}
-        valid={fieldTouched ? (noTickShow && !fieldError ? null : !fieldError) : null}
+        valid={fieldTouched && !readOnly ? (noTickShow && !fieldError ? null : !fieldError) : null}
+        readOnly={readOnly || false}
       />
+      {fieldError && fieldTouched ? <ErrorMessage>{fieldError}</ErrorMessage> : <ErrorMessagePleaceHolder />}
     </Wrapper>
   );
 };
 
-export default WorldField;
+export default TextField;