+++ app/renderer/pages/wallet/account/AccountDetails.js
@@ -1,36 +1,52 @@
 import React from 'react';
-import R from 'ramda';
-import styled from 'styled-components';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { MainContent, MainLayout } from 'components/layout';
+
+import { CENNZScanAddressUrl } from 'common/types/cennznet-node.types';
 import { Clipboard, PageHeading, PageFooter, Tabs, TabPane } from 'components';
 import PortfolioSection from './PortfolioSection';
 import ReceiveSection from './ReceiveSection';
+import TransferSection from './transferSectionPage';
+import ClipboardShareLinks from './transferSectionPage/ClipboardShareLinks';
 
 const Subheading = ({ account }) => {
+  const url = CENNZScanAddressUrl.rimu; // TODO should base on selected network
   return (
     <div style={{ display: 'flex' }}>
       <span>Public Address: </span>
-      <Clipboard>{account.address}</Clipboard>
+      <ClipboardShareLinks
+        url={url}
+        styles={{
+          height: '1rem',
+          lineHeight: '1rem',
+          padding: null,
+          backgroundColor: null,
+          textDecoration: null,
+          icon2MarginLeft: '1rem',
+          textPaddingTop: null,
+          textMinWidth: null,
+        }}>{account.address}
+      </ClipboardShareLinks>
     </div>
   );
 };
 
-const AccountDetails = ({ account }) => {
+const AccountDetails = ({ account, onTransfer, currentWallet, transaction }) => {
   return account ? (
     <React.Fragment>
       <PageHeading subHeading={<Subheading {...{ account }} />}>Account name here...</PageHeading>
       <div className="content">
-        <Tabs defaultActiveKey="2" styles={{ maxHeight: '730px' }}>
-          <TabPane tab="Portfolio" key="1">
+        <Tabs defaultActiveKey="1" >
+          <TabPane tab="Portfolio" key="1" withScrollable>
             <PortfolioSection {...{ account }} />
           </TabPane>
-          <TabPane tab="Receive" key="2">
+          <TabPane tab="Deposit" key="2" withScrollable>
             <ReceiveSection {...{ account }} />
+          </TabPane>
+          <TabPane tab="Send" key="3" >
+            <TransferSection {...{ account, onTransfer, currentWallet, transaction }} />
           </TabPane>
         </Tabs>
       </div>
-      <PageFooter blank />
     </React.Fragment>
   ) : (
     <PageHeading subHeading="There is no accessible account"> Wallet account </PageHeading>