+++ app/renderer/pages/wallet/account/container.js
@@ -1,7 +1,6 @@
 import { connect } from 'react-redux';
-import { compose, lifecycle } from 'recompose';
+import { compose, lifecycle, withState } from 'recompose';
 import R from 'ramda';
-
 
 import types from 'renderer/types';
 import { storageKeys } from 'renderer/api/utils/storage';
@@ -9,13 +8,21 @@ import history from 'renderer/history';
 import ROUTES from 'renderer/constants/routes';
 import { Logger } from 'renderer/utils/logging';
 
-const mapStateToProps = ({ localStorage }) => ({ wallets: localStorage[storageKeys.WALLETS] });
+const mapStateToProps = ({ localStorage, transaction }) => ({
+  wallets: localStorage[storageKeys.WALLETS],
+  transaction,
+});
 
 const mapDispatchToProps = dispatch => ({
   onPageLoaded: payload => {},
   onSyncWalletData: payload => {
     Logger.debug('onSyncWalletData');
     dispatch({ type: types.syncWalletData.requested, payload });
+  },
+
+  onTransfer: payload => {
+    Logger.debug('onTransfer');
+    dispatch({ type: types.transfer.requested, payload });
   },
 });
 
@@ -33,7 +40,6 @@ const enhance = compose(
     },
 
     componentDidUpdate(prevProps) {
-
       // sync wallet data when nav to different account
       if (this.props.match.params.walletId !== prevProps.match.params.walletId) {
         Logger.debug('sync wallet data on different wallet Id');
@@ -42,7 +48,7 @@ const enhance = compose(
         const wallet = R.find(R.propEq('id', walletId))(wallets);
         onSyncWalletData({ id: walletId, wallet });
       }
-    }
+    },
   })
 );