+++ app/renderer/pages/wallet/create/container.js
@@ -15,10 +15,15 @@ const mapStateToProps = ({
 
 const mapDispatchToProps = dispatch => ({
   onCreateWallet: payload => {
-    dispatch({ type: types.walletCreatWithSKR.requested, payload });
+    return window.odin.api.cennz.createWalletWithHDKeyRing(payload);
   },
+
   onCreatePaperWallet: payload => {
     dispatch({ type: types.walletPaperGenerate.requested, payload });
+  },
+
+  onCompleteWalletCreation: payload => {
+    dispatch({ type: types.walletCreatWithHDKR.requested, payload });
   },
 });
 
@@ -34,17 +39,23 @@ const enhance = compose(
       initMnemonic = '',
       initWalletName = '',
       initIsOpenPenPrepareModal = false,
+      initError = null,
+      initWallet = null,
     }) => ({
       step: initStep,
       mnemonicString: initMnemonic,
       walletName: initWalletName,
       isOpenPenPrepareModal: initIsOpenPenPrepareModal,
+      error: initError,
+      wallet: initWallet,
     }),
     {
       moveToStep: () => val => ({ step: val }),
       setMnemonicString: () => val => ({ mnemonicString: val }),
       setWalletName: () => val => ({ walletName: val }),
       setIsOpenPenPrepareModal: () => val => ({ isOpenPenPrepareModal: val }),
+      setError: () => val => ({ error: val }),
+      setWallet: () => val => ({ wallet: val }),
     }
   )
 );