+++ CHANGELOG.md
@@ -3,19 +3,42 @@
 # vNext
 
 ### Features
+- add CENNZScan address link to account public address
 
 ### Fixes
 
 ### Improvements
+- tidy up asset table layout
+- update background gradient settings
+- update field text bg color to match with design
+- update sync info layout
+
+- Refactor configure
+
+# v0.5.0 (2019.02.20)
+
+### Features
+
+- able to send tx
+- add Yup validations for send tx
+
+### Fixes
+
+- fix switch network issues
+
+### Improvements
+
+- Make 'Local test net' optional
 
 # v0.4.1 (2019.02.12)
 
 ### Features
 
-- Add tabs and scrallable components
+- Add tabs and scrollable components
 - new data structure to store Wallets -> Accounts -> Assets
 - Connect wallet with both HDKeyring and SimpleKeyring
 - Able to display balance in Account detail page
+- Add send tx confirmation flow
 
 ### Fixes